19-deepeval/requirements.txt
google-adk>=1.19.0
deepeval>=3.7.2
python-dotenv>=1.2.1
